cambio en models clase abstract, en services agrege interfaces

+++ src/main/java/com/faz/sadminpasajes/repositorys/ChoferRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChoferRepository extends JpaRepository<Chofer,Integer> {
+public interface ChoferRepository extends JpaRepository<Chofer,Long> {
     
     Optional<Chofer> findByDni(int dni);
 }
