+++ src/main/java/com/faz/sadminpasajes/repositorys/EmpresaRepository.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpresaRepository extends JpaRepository<Empresa,Integer> {
+public interface EmpresaRepository extends JpaRepository<Empresa,Long> {
     Optional<Empresa> findByNombre(String nom);
 }
